Refactored ML code, more readable. Made the code packaging better and cleaner.
