+++ data/suffix_vocab.json
@@ -1,0 +1,131 @@
+{
+  "suffix_to_id": {
+    "reflexive_is": 0,
+    "reflexive_ik": 1,
+    "active_t": 2,
+    "active_tir": 3,
+    "active_ir": 4,
+    "passive_il": 5,
+    "passive_in": 6,
+    "negative_me": 7,
+    "dative_archaic_ke": 8,
+    "locative_den": 9,
+    "dative_e": 10,
+    "ablative_de": 11,
+    "plural_ler": 12,
+    "counting_er": 13,
+    "cooperative_daş": 14,
+    "relative_ce": 15,
+    "relative_sel": 16,
+    "dimunitive_ek_archaic": 17,
+    "approximative_imtrak": 18,
+    "accusative": 19,
+    "pluralizer_archaic_iz": 20,
+    "genitive_im": 21,
+    "genitive_in": 22,
+    "dimunitive_cik": 23,
+    "actor_ci": 24,
+    "ordinal_inci": 25,
+    "approximative_si": 26,
+    "privative_siz": 27,
+    "philicative_cil": 28,
+    "composessive_li": 29,
+    "suitative_lik": 30,
+    "adjectiative_ay": 60,
+    "absentative_se": 32,
+    "verbifier_e": 33,
+    "aplicative_le": 34,
+    "grantative_let": 35,
+    "transformative_les": 36,
+    "acquirative_len": 37,
+    "verbifier_ik": 38,
+    "nounifier_ecek": 39,
+    "infinitive_ma": 40,
+    "infinitive_mak": 41,
+    "factative_en": 42,
+    "factative_er": 43,
+    "toolative_en": 44,
+    "adverbial_e": 45,
+    "constofactative_gen": 46,
+    "constofactative_gin": 47,
+    "nounifier_iş": 48,
+    "perfectative_ik": 49,
+    "nounifier_i": 50,
+    "nounifier_gi": 51,
+    "nounifier_im": 52,
+    "nounifier_in": 53,
+    "nounifier_inç": 54,
+    "nounifier_inti": 55,
+    "toolifier_geç": 56,
+    "toolifier_eç": 57,
+    "neverfactative_mez": 58,
+    "pastfactative_miş": 59
+  },
+  "id_to_suffix": {
+    "0": "reflexive_is",
+    "1": "reflexive_ik",
+    "2": "active_t",
+    "3": "active_tir",
+    "4": "active_ir",
+    "5": "passive_il",
+    "6": "passive_in",
+    "7": "negative_me",
+    "8": "dative_archaic_ke",
+    "9": "locative_den",
+    "10": "dative_e",
+    "11": "ablative_de",
+    "12": "plural_ler",
+    "13": "counting_er",
+    "14": "cooperative_daş",
+    "15": "relative_ce",
+    "16": "relative_sel",
+    "17": "dimunitive_ek_archaic",
+    "18": "approximative_imtrak",
+    "19": "accusative",
+    "20": "pluralizer_archaic_iz",
+    "21": "genitive_im",
+    "22": "genitive_in",
+    "23": "dimunitive_cik",
+    "24": "actor_ci",
+    "25": "ordinal_inci",
+    "26": "approximative_si",
+    "27": "privative_siz",
+    "28": "philicative_cil",
+    "29": "composessive_li",
+    "30": "suitative_lik",
+    "31": "adjectiative_ay",
+    "32": "absentative_se",
+    "33": "verbifier_e",
+    "34": "aplicative_le",
+    "35": "grantative_let",
+    "36": "transformative_les",
+    "37": "acquirative_len",
+    "38": "verbifier_ik",
+    "39": "nounifier_ecek",
+    "40": "infinitive_ma",
+    "41": "infinitive_mak",
+    "42": "factative_en",
+    "43": "factative_er",
+    "44": "toolative_en",
+    "45": "adverbial_e",
+    "46": "constofactative_gen",
+    "47": "constofactative_gin",
+    "48": "nounifier_iş",
+    "49": "perfectative_ik",
+    "50": "nounifier_i",
+    "51": "nounifier_gi",
+    "52": "nounifier_im",
+    "53": "nounifier_in",
+    "54": "nounifier_inç",
+    "55": "nounifier_inti",
+    "56": "toolifier_geç",
+    "57": "toolifier_eç",
+    "58": "neverfactative_mez",
+    "59": "pastfactative_miş",
+    "60": "adjectiative_ay"
+  },
+  "category_to_id": {
+    "Noun": 0,
+    "Verb": 1
+  }
+}
