--- old_versions/old_suffixes/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ['name2name' , 'verb2name', 'verb2verb', 'name2verb']
