--- old_versions/old_suffixes/name2name.py
@@ -1,100 +1,0 @@
-
-#name2name_
-back = [
-    'lar',
-    "lık",
-    "luk",
-    'lı'  ,
-    'lu'  ,
-    'cı'  ,
-    'cu'  ,
-    'çı'  ,
-    'çu'  ,
-    'cık'  ,
-    'cuk'  ,
-    'çık'  ,
-    'çuk'  ,   
-    'ca'  ,
-    'ça'  ,
-    'sı'  ,
-    'su'  ,
-    'umsu',
-    'ımsı',
-    'ımtrak',
-    'umtrak',
-    'cıl'   ,
-    'cul'   ,
-    'çıl'   ,
-    'çul'   ,
-    'sız'   ,
-    'suz'   ,    
-    'daş'   , 
-
-    'ar' , #ulestirme
-    'ıncı',
-    'uncu',
-
-]
-
-
-#name2name_
-front = [
-    'ler',
-    'lik' ,#luk ve turevler          kucuk unlu uyumu
-    "lük",
-
-    'li'  ,# LI ve LU vs             kucuk unlu uyumu
-
-    'lü'  ,
-    
-    
-    
-    'ci'  ,#cu ve çu varyantlari    kucuk unlu uyumu
-    'cü'  ,
-
-    'çi'  ,
-
-    'çü'  ,
-    
-    'cik'  ,#cu ve çu varyantlari    kucuk unlu uyumu
-
-    'cük'  ,
-
-    'çik'  ,
-
-    'çük'  ,
-    
-    'ce'  ,#ca ve cha                
-
-    'çe'  ,
-
-    
-    'si'  ,#sI                       kucuk unlu uyumu
-    'sü'  ,
-
-
-    'imsi',#ImsI                     kucuk unlu uyumu
-    'ümsü',
-
-    'imtrak',#Imtrak                 kucuk unlu uyumu
-
-    'ümtrak',
-    
-    'cil'   , # CUL vs varyantlar    kucuk unlu uyumu
-
-    'cül'   ,
-
-    'çil'   ,
-
-    'çül'   ,
-
-
-
-    'siz'   , #sIz SUZ seuz vs       kucuk unlu uyumu 
-    'süz'   ,    
-    
-    'deş'   ,
-    'er', #ulestirme
-    'inci', #sıra
-    'üncü'
-]
