--- old_versions/old_suffixes/name2verb.py
@@ -1,30 +1,0 @@
-#name2verb_
-back = [
-
-    'la',
-    'a',
-    'laş',
-    'ık',
-    'lan',
-    'sa' ,
-    'ar',
-    'ır',
-    "lat",
-]
-
-#name2verb_
-front = [
-
-    'le',
-    'e',
-    'leş',
-    'ik',
-    'len',
-    "let",
-    'se' ,
-    'er',
-    'ir',
-]
-
-
-
