--- old_versions/old_suffixes/verb2name.py
@@ -1,52 +1,0 @@
-#verb2name_
-back = [
-    "a",
-    "gı",
-    "gu",
-    "ım",
-    "um",
-    "ık",
-    "uk",
-    "ak",
-    "ın",
-    "un",
-    "gın",
-    "gun",
-    "gan",
-    "kan",
-    "kun",
-    "kın",
-    "ınç",
-    "mak",
-    "ma",
-    "u",
-    "ı",
-    "an",
-
-]
-
-#verb2name_
-front = [
-    "e",
-    "gi",
-    "gü",
-    "im",
-    "üm",
-    "ik",
-    "ük",
-    "ek",
-    "in",
-    "ün",
-    "gin",
-    "gün",
-    "gen",
-    "ken",
-    "kün",
-    "kin",
-    "inç",
-    "mek",
-    "me",
-    "ü",
-    "i",
-    "en",
-]
