--- old_versions/old_suffixes/verb2verb.py
@@ -1,42 +1,0 @@
-#verb2verb_
-back = [
-    "ma",
-    'ıt',
-    "ut",
-    't',
-    'dır', 
-    'dur',
-    'tır',
-    'tur',
-    'ış',
-    "uş",
-    'ş',
-    'ın',
-    'un',
-    'n',
-    'ık',
-    'ıl',
-    'l',
-
-]
-
-#verb2verb_
-front = [
-    "me",
-    'it',
-    "üt",
-    't',
-    'dir', 
-    'dür',
-    'tir',
-    'tur',
-    'iş',
-    "üş",
-    'ş',
-    'in',
-    'ün',
-    'n',
-    'ik',
-    'il',
-    'l',
-]
