--- suffix_vocab.json
@@ -1,128 +1,0 @@
-{
-  "suffix_to_id": {
-    "reflexive_is": 0,
-    "reflexive_ik": 1,
-    "active_t": 2,
-    "active_tir": 3,
-    "active_ir": 4,
-    "passive_il": 5,
-    "passive_in": 6,
-    "negative_me": 7,
-    "dative_archaic_ke": 8,
-    "locative_den": 9,
-    "dative_e": 10,
-    "ablative_de": 11,
-    "plural_ler": 12,
-    "counting_er": 13,
-    "cooperative_daş": 14,
-    "relative_ce": 15,
-    "relative_sel": 16,
-    "dimunitive_ek_archaic": 17,
-    "approximative_imtrak": 18,
-    "accusative": 19,
-    "pluralizer_archaic_iz": 20,
-    "genitive_im": 21,
-    "genitive_in": 22,
-    "dimunitive_cik": 23,
-    "actor_ci": 24,
-    "ordinal_inci": 25,
-    "approximative_si": 26,
-    "privative_siz": 27,
-    "philicative_cil": 28,
-    "composessive_li": 29,
-    "suitative_lik": 30,
-    "absentative_se": 31,
-    "verbifier_e": 32,
-    "aplicative_le": 33,
-    "grantative_let": 34,
-    "transformative_les": 35,
-    "acquirative_len": 36,
-    "verbifier_ik": 37,
-    "nounifier_ecek": 38,
-    "infinitive_ma": 39,
-    "infinitive_mak": 40,
-    "factative_en": 41,
-    "factative_er": 42,
-    "toolative_en": 43,
-    "adverbial_e": 44,
-    "constofactative_gen": 45,
-    "constofactative_gin": 46,
-    "nounifier_iş": 47,
-    "perfectative_ik": 48,
-    "nounifier_i": 49,
-    "nounifier_gi": 50,
-    "nounifier_im": 51,
-    "nounifier_in": 52,
-    "nounifier_inç": 53,
-    "nounifier_inti": 54,
-    "toolifier_geç": 55,
-    "toolifier_eç": 56,
-    "neverfactative_mez": 57,
-    "pastfactative_miş": 58
-  },
-  "id_to_suffix": {
-    "0": "reflexive_is",
-    "1": "reflexive_ik",
-    "2": "active_t",
-    "3": "active_tir",
-    "4": "active_ir",
-    "5": "passive_il",
-    "6": "passive_in",
-    "7": "negative_me",
-    "8": "dative_archaic_ke",
-    "9": "locative_den",
-    "10": "dative_e",
-    "11": "ablative_de",
-    "12": "plural_ler",
-    "13": "counting_er",
-    "14": "cooperative_daş",
-    "15": "relative_ce",
-    "16": "relative_sel",
-    "17": "dimunitive_ek_archaic",
-    "18": "approximative_imtrak",
-    "19": "accusative",
-    "20": "pluralizer_archaic_iz",
-    "21": "genitive_im",
-    "22": "genitive_in",
-    "23": "dimunitive_cik",
-    "24": "actor_ci",
-    "25": "ordinal_inci",
-    "26": "approximative_si",
-    "27": "privative_siz",
-    "28": "philicative_cil",
-    "29": "composessive_li",
-    "30": "suitative_lik",
-    "31": "absentative_se",
-    "32": "verbifier_e",
-    "33": "aplicative_le",
-    "34": "grantative_let",
-    "35": "transformative_les",
-    "36": "acquirative_len",
-    "37": "verbifier_ik",
-    "38": "nounifier_ecek",
-    "39": "infinitive_ma",
-    "40": "infinitive_mak",
-    "41": "factative_en",
-    "42": "factative_er",
-    "43": "toolative_en",
-    "44": "adverbial_e",
-    "45": "constofactative_gen",
-    "46": "constofactative_gin",
-    "47": "nounifier_iş",
-    "48": "perfectative_ik",
-    "49": "nounifier_i",
-    "50": "nounifier_gi",
-    "51": "nounifier_im",
-    "52": "nounifier_in",
-    "53": "nounifier_inç",
-    "54": "nounifier_inti",
-    "55": "toolifier_geç",
-    "56": "toolifier_eç",
-    "57": "neverfactative_mez",
-    "58": "pastfactative_miş"
-  },
-  "category_to_id": {
-    "Noun": 0,
-    "Verb": 1
-  }
-}
