--- training_count.txt
@@ -1,1 +1,0 @@
-154
